Rename some classes for PHP7 compatibility with strict type hints v5
