+++ tests/Mutator/Number/FloatValueTest.php
@@ -11,14 +11,14 @@
 
 namespace Humbug\Test\Mutator\Number;
 
-use Humbug\Mutator;
+use Humbug\Mutator\Number\FloatValue;
 
-class FloatTest extends \PHPUnit_Framework_TestCase
+class FloatValueTest extends \PHPUnit_Framework_TestCase
 {
 
     public function testReturnsTokenSwitchingZeroForOne()
     {
-        $mutation = new Mutator\Number\Float;
+        $mutation = new FloatValue;
         $tokens = [10 => [T_DNUMBER, 0.0]];
         $mutation->getMutation($tokens, 10);
         $this->assertEquals([10 => [T_DNUMBER, 1.0]], $tokens);
@@ -26,7 +26,7 @@ class FloatTest extends \PHPUnit_Framework_TestCase
 
     public function testReturnsTokenSwitchingOneForZero()
     {
-        $mutation = new Mutator\Number\Float;
+        $mutation = new FloatValue;
         $tokens = [10 => [T_DNUMBER, 1.0]];
         $mutation->getMutation($tokens, 10);
         $this->assertEquals([10 => [T_DNUMBER, 0.0]], $tokens);
@@ -34,7 +34,7 @@ class FloatTest extends \PHPUnit_Framework_TestCase
 
     public function testReturnsTokenSwitchingBetweenOneAndTwoWithIncrement()
     {
-        $mutation = new Mutator\Number\Float;
+        $mutation = new FloatValue;
         $tokens = [10 => [T_DNUMBER, 1.5]];
         $mutation->getMutation($tokens, 10);
         $this->assertEquals([10 => [T_DNUMBER, 2.5]], $tokens);
@@ -42,7 +42,7 @@ class FloatTest extends \PHPUnit_Framework_TestCase
 
     public function testReturnsTokenSwitchingAnyIntegerGreaterThanTwoWithOne()
     {
-        $mutation = new Mutator\Number\Float;
+        $mutation = new FloatValue;
         $tokens = [10 => [T_DNUMBER, 2.0]];
         $mutation->getMutation($tokens, 10);
         $this->assertEquals([10 => [T_DNUMBER, 1.0]], $tokens);
@@ -51,18 +51,18 @@ class FloatTest extends \PHPUnit_Framework_TestCase
     public function testMutatesFloat()
     {
         $tokens = [10 => [T_DNUMBER, 1.05]];
-        $this->assertTrue(Mutator\Number\Float::mutates($tokens, 10));
+        $this->assertTrue(FloatValue::mutates($tokens, 10));
     }
 
     public function testDoesNotMutateOtherScalars()
     {
         $tokens = [11 => [T_STRING, 'foo']];
-        $this->assertFalse(Mutator\Number\Float::mutates($tokens, 11));
+        $this->assertFalse(FloatValue::mutates($tokens, 11));
 
         $tokens = [11 => [T_LNUMBER, 0]];
-        $this->assertFalse(Mutator\Number\Float::mutates($tokens, 11));
+        $this->assertFalse(FloatValue::mutates($tokens, 11));
 
         $tokens = [11 => [T_LNUMBER, 1]];
-        $this->assertFalse(Mutator\Number\Float::mutates($tokens, 11));
+        $this->assertFalse(FloatValue::mutates($tokens, 11));
     }
 }
